+++ CHANGELOG.md
@@ -2,7 +2,25 @@
 
 <!--next-version-placeholder-->
 
-## v0.2.1 (_/02/2024)
+## v0.2.3 (07/03/2024)
+- New feature: `bayesbay.prior.LaplacePrior`
+- New feature: `bayesbay.LogLikelihood.__repr__`
+- New feature: `bayesbay.Target.__repr__`
+
+## v0.2.2 (04/03/2024)
+- New feature: `Voronoi2D`
+    - `Voronoi2D.get_tessellation_statistics()` added
+- API change: `Voronoi1D`
+    - generalized all static methods:
+        - interpolate_tessellation
+        - _interpolate_tessellations
+        - get_tessellation_statistics
+        - plot_tessellation
+        - plot_tessellations
+        - plot_tessellation_statistics
+        - plot_tessellation_density
+
+## v0.2.1 (27/02/2024)
 - New feature: `Voronoi2D`
     - `Voronoi2D.interpolate_tessellation()` added
     - `Voronoi2D.polygon` added, allowing for prior probabilities defined within polygons